Extracted `to_sql()` to SqlDust.QueryUtils
